Сomplete homework 7
