+++ main.js
@@ -1,78 +1,6 @@
-function createElement(tag, className) {
-    const $tag = document.createElement(tag);
-    if (className) {
-        $tag.classList.add(className);
-    }
-
-    return $tag;
-}
-
-function getRandom(max) {
-    const random = Math.floor(Math.random() * max);
-    return random;
-}
-
-function changeHP(hp) {
-    this.hp -= hp;
-
-    if (this.hp <= 0) {
-        this.hp = 0;
-    }
-};
-
-function elHP() {
-    return document.querySelector(`.player${this.player} .life`)
-};
-
-function renderHP() {
-    const $playerLife = this.elHP();
-    $playerLife.style.width = `${this.hp}%`;
-};
-
-class Player {
-    constructor(characterData) {
-        this.player = characterData.player;
-        this.name = characterData.name;
-        this.hp = characterData.hp;
-        this.img = characterData.img;
-        this.weapon = characterData.weapon;
-
-        this.changeHP = changeHP;
-        this.elHP = elHP;
-        this.renderHP = renderHP;
-    }
-
-    attack() {
-        console.log(`${this.name} Fight...`)
-    }
-
-    createPlayer() {
-        const $player = createElement("div", `player${this.player}`);
-
-        const $progressbar = createElement("div", "progressbar");
-
-        const $life = createElement("div", "life");
-        $life.style.width = `${this.hp}%`;
-
-        const $name = createElement("div", "name");
-        $name.innerText = this.name;
-
-        const $character = createElement("div", "character");
-
-        const $img = createElement("img", "character");
-        $img.src = this.img;
-
-        $progressbar.appendChild($life);
-        $progressbar.appendChild($name);
-
-        $character.appendChild($img);
-
-        $player.appendChild($progressbar);
-        $player.appendChild($character);
-
-        return $player;
-    }
-}
+import { createElement, getRandom } from "./utils.js";
+import Player from "./player.js";
+import { scorpionData, subzeroData, HIT, ATTACK, logs } from "./data.js";
 
 class Game {
     constructor(player1Data, player2Data) {
@@ -229,69 +157,6 @@ class Game {
         this.player2.attack();
     }
 }
-
-let scorpionData = {
-    player: 1,
-    name: "SCORPION",
-    hp: 100,
-    img: "http://reactmarathon-api.herokuapp.com/assets/scorpion.gif",
-    weapon: ["HOOK"]
-}
-
-let subzeroData = {
-    player: 2,
-    name: "SUB-ZERO",
-    hp: 100,
-    img: "http://reactmarathon-api.herokuapp.com/assets/subzero.gif",
-    weapon: ["SWORD"]
-}
-
-const HIT = {
-    head: 30,
-    body: 25,
-    foot: 20,
-}
-const ATTACK = ['head', 'body', 'foot'];
-
-const logs = {
-    start: 'Часы показывали [time], когда [player1] и [player2] бросили вызов друг другу.',
-    end: [
-        'Результат удара [playerWins]: [playerLose] - труп',
-        '[playerLose] погиб от удара бойца [playerWins]',
-        'Результат боя: [playerLose] - жертва, [playerWins] - убийца',
-    ],
-    hit: [
-        '[playerDefence] пытался сконцентрироваться, но [playerKick] разбежавшись раздробил копчиком левое ухо врага.',
-        '[playerDefence] расстроился, как вдруг, неожиданно [playerKick] случайно раздробил грудью грудину противника.',
-        '[playerDefence] зажмурился, а в это время [playerKick], прослезившись, раздробил кулаком пах оппонента.',
-        '[playerDefence] чесал <вырезано цензурой>, и внезапно неустрашимый [playerKick] отчаянно размозжил грудью левый бицепс оппонента.',
-        '[playerDefence] задумался, но внезапно [playerKick] случайно влепил грубый удар копчиком в пояс оппонента.',
-        '[playerDefence] ковырялся в зубах, но [playerKick] проснувшись влепил тяжелый удар пальцем в кадык врага.',
-        '[playerDefence] вспомнил что-то важное, но внезапно [playerKick] зевнув, размозжил открытой ладонью челюсть противника.',
-        '[playerDefence] осмотрелся, и в это время [playerKick] мимоходом раздробил стопой аппендикс соперника.',
-        '[playerDefence] кашлянул, но внезапно [playerKick] показав палец, размозжил пальцем грудь соперника.',
-        '[playerDefence] пытался что-то сказать, а жестокий [playerKick] проснувшись размозжил копчиком левую ногу противника.',
-        '[playerDefence] забылся, как внезапно безумный [playerKick] со скуки, влепил удар коленом в левый бок соперника.',
-        '[playerDefence] поперхнулся, а за это [playerKick] мимоходом раздробил коленом висок врага.',
-        '[playerDefence] расстроился, а в это время наглый [playerKick] пошатнувшись размозжил копчиком губы оппонента.',
-        '[playerDefence] осмотрелся, но внезапно [playerKick] робко размозжил коленом левый глаз противника.',
-        '[playerDefence] осмотрелся, а [playerKick] вломил дробящий удар плечом, пробив блок, куда обычно не бьют оппонента.',
-        '[playerDefence] ковырялся в зубах, как вдруг, неожиданно [playerKick] отчаянно размозжил плечом мышцы пресса оппонента.',
-        '[playerDefence] пришел в себя, и в это время [playerKick] провел разбивающий удар кистью руки, пробив блок, в голень противника.',
-        '[playerDefence] пошатнулся, а в это время [playerKick] хихикая влепил грубый удар открытой ладонью по бедрам врага.',
-    ],
-    defence: [
-        '[playerKick] потерял момент и храбрый [playerDefence] отпрыгнул от удара открытой ладонью в ключицу.',
-        '[playerKick] не контролировал ситуацию, и потому [playerDefence] поставил блок на удар пяткой в правую грудь.',
-        '[playerKick] потерял момент и [playerDefence] поставил блок на удар коленом по селезенке.',
-        '[playerKick] поскользнулся и задумчивый [playerDefence] поставил блок на тычок головой в бровь.',
-        '[playerKick] старался провести удар, но непобедимый [playerDefence] ушел в сторону от удара копчиком прямо в пятку.',
-        '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.',
-        '[playerKick] не думал о бое, потому расстроенный [playerDefence] отпрыгнул от удара кулаком куда обычно не бьют.',
-        '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.'
-    ],
-    draw: 'Ничья - это тоже победа!'
-};
 
 window.onload = () => {
     const game = new Game(scorpionData, subzeroData);
